#101 Move AutoValueEditor into variable control.
